+++ chainerx_cc/chainerx/cuda/cuda_device.h
@@ -22,6 +22,7 @@
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
 #include "chainerx/kernels/normalization.h"
+#include "chainerx/kernels/pooling.h"
 #include "chainerx/routines/normalization.h"
 #include "chainerx/routines/pooling.h"
 #include "chainerx/scalar.h"
