+++ chainerx_cc/chainerx/cuda/cuda_device/pool.cu
@@ -23,6 +23,7 @@
 #include "chainerx/error.h"
 #include "chainerx/indexable_array.h"
 #include "chainerx/indexer.h"
+#include "chainerx/kernels/pooling.h"
 #include "chainerx/macro.h"
 #include "chainerx/numeric_limits.h"
 #include "chainerx/routines/connection.h"
